Revise pipeline scripts for Nextflow integration

+++ scripts/synteny_search/intersect_blast_and_synteny_pairwiseBlastApproach.sh
@@ -18,13 +18,15 @@
 
 
 print_help() {
-    echo "Usage: $0 -b <blast> -s <synteny> -p <pident> -c<'genome'|'locus'> -o <outname> -h"
-    echo "        Outputs: to $outname, writes the output of the pairwise BLAST of $synteny as query against $blast as subject."
+    echo "Usage: $0 -b <blast> -s <synteny> -m <multiple> -p <pident> -q <qcovs> -c<'genome'|'locus'> -o <outname> -h"
+    echo "        Outputs: to -o, writes the output of the pairwise BLAST of $synteny as query against $blast as subject."
     echo "                 Rationale for this direction of BLAST is explained in comments at the top of this script."
-    echo "                 Also writes a fasta with $pident cutoff- look for a file with a name similar to outname."
+    echo "                 Also writes a fasta with -p and -q cutoff- look for a file with a name similar to outname."
     echo "  -b    Required. Path to BLAST file in which the first five columns are sgi sseqid sseq evalue stitle."
     echo "  -s    Required. Path to summary file from synteny search via synteny_wrapper.sh (which calls find_synteny_hits.sh)."
-    echo "  -p    Optional. Pident threshold for filtering (default = 99)."
+    echo "  -m    Optional. If true, treats the -s input as a file referring to multiple outdirs and thus multiple summary files (default false)."
+    echo "  -p    Optional. Pident threshold for filtering, discarding any hits with pident lower than this (default = 99)."
+    echo "  -q    Optional. qcovs threshold for filtering, discarding any hits with pident lower than this (default = 90)."
     echo "  -c    Optional. Concatenation mode for convert_blast_to_fasta.sh (default = 'no')."
     echo "  -k    Optional. Keep FASTA of all sequences in $synteny (default = 'false')."
     echo "  -o    Required. Name of output file."
@@ -32,16 +34,20 @@ print_help() {
 }
 
 # set default values for optional arguments
+multiple="false"
 pident="99"
+qcovs="90"
 concat="no"
 keep="false"
 
 # Parse options
-while getopts "b:s:p:c:ko:h" opt; do
+while getopts "b:s:mp:q:c:ko:h" opt; do
   case $opt in
     b) blast="$OPTARG" ;;
     s) synteny="$OPTARG" ;;
+    m) multiple="true" ;;
     p) pident="$OPTARG" ;;
+    q) qcovs="$OPTARG" ;;
     c) concat="$OPTARG" ;;
     k) keep="true" ;;
     o) outname="$OPTARG" ;;
@@ -65,6 +71,7 @@ echo "Saving tempBlastFasta to ${tempBlastFasta}..."
 
 # remove gap characters "-" because command line BLAST doesn't recognize them.
 # only write a line if the sequence hasn't been encountered before; this ensures the FASTA only contains unique seqs
+# remove gaps BEFORE checking for uniqueness- sequences with gaps may be identical without gaps
 awk 'BEGIN {OFS="\t"} {
   gsub("-", "", $3);
   if (!seen[$3]++) print
@@ -98,11 +105,13 @@ yes | rm $tempSynteny
 temp_dir=$(mktemp -d)
 echo "Temporary directory for blast db created at: $temp_dir"
 mkdir -p "${temp_dir}/temp_db"
+currDir=$pwd # save this; especially if running this script in Nextflow, need to be careful about current directory location
 cd $temp_dir # need to be one directory above the blast db being made
 makeblastdb -in "$tempBlastFasta" -input_type fasta -dbtype prot -title temp_db -out "${temp_dir}/temp_db" &
 pid3=$!
 wait $pid3
 yes | rm -f $tempBlastFasta
+cd $currDir
 
 ### Perform the pairwise BLAST
 # conda deactivate # currently, local blast is in the base conda environment away from blast_env
@@ -112,27 +121,34 @@ echo "Saving pairwise blast outputs..."
 tempBlastOutputs=$(mktemp) # going to be using awk to add a header later, so for now I'm saving it to a temp file
 procs_to_use=$(( $(nproc) / 4 ))
 max=10
-blastp -query "$tempSyntenyFasta" -db "${temp_dir}/temp_db" -max_target_seqs "$max" -num_threads "$procs_to_use" -outfmt "6 qseqid qseq sallseqid sseq pident evalue bitscore qcovs" -out "$tempBlastOutputs" &
+blastp -query "$tempSyntenyFasta" -db "${temp_dir}/temp_db" -max_target_seqs "$max" -num_threads "$procs_to_use" -outfmt "6 qseqid qseq sallseqid sseq pident qcovs evalue bitscore" -out "$tempBlastOutputs" &
 pid4=$!
 
 wait $pid4
 
 ### Delete the blast db that was created
 #cleanup-blastdb-volumes.py -db temp_db -dbtype prot
-rm -rf "$temp_dir"
+if [[ $keep = "true" ]]; then
+  blastDBname="${blast%.*}_db"
+  mv $temp_dir $blastDBname
+  echo "Saved blast DB made from input BLAST file to $blastDBname"
+else
+  rm -rf "$temp_dir"
+fi
+
 
 ### Filter the output file from the pairwise blast to only include lines with >=$pident
 tempFilteredIDs=$(mktemp)
-awk -v p="$pident" '$5 >= p' "$tempBlastOutputs" | cut -f 1 | sort | uniq > "$tempFilteredIDs" # IDs were in column 1 of the blast that was just done, pident was in column 5
+awk -v p="$pident" -v q="$qcovs" '$5 >= p && $6 >= q' "$tempBlastOutputs" | cut -f 1 | sort | uniq > "$tempFilteredIDs" # IDs were in column 1 of the blast that was just done, pident was in column 5
 
 ### Filter the synteny fasta to only include the lines from the filtered IDs list
 # This filtered FASTA is the final output of this script. You can still left join the synteny metadata to the seq IDs in that filtered FASTA; it's a subset.
-fastaname="${outname%.*}_pident${pident}.fasta"
+fastaname="${outname%.*}_pident${pident}_qcovs${qcovs}.fasta"
 seqtk subseq "$tempSyntenyFasta" "$tempFilteredIDs" > "$fastaname"
 echo "Saved filtered FASTA to $fastaname"
 
 # Add colnames to the file saved to $outname
-awk 'BEGIN { OFS="\t"; print "qseqid", "qseq", "sallseqid", "sseq", "pident", "evalue", "bitscore" } { print }' "$tempBlastOutputs" > "$outname"
+awk 'BEGIN { OFS="\t"; print "qseqid", "qseq", "sallseqid", "sseq", "pident", "qcovs", "evalue", "bitscore" } { print }' "$tempBlastOutputs" > "$outname"
 
 if [[ $keep = "true" ]]; then
   syntenyFastaName="${synteny%.*}.fasta"
